LayoutTests/fast/events/ios/non-meaningful-click-when-tapping-document.html

<!DOCTYPE html> <!-- webkit-test-runner [ useFlexibleViewport=true ] -->
<html>
<head>
<meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=no">
<style>
body, html {
    width: 100%;
    height: 100%;
}
</style>
<script src="../../../resources/js-test.js"></script>
<script src="../../../resources/ui-helper.js"></script>
<script>
jsTestIsAsync = true;

addEventListener("load", async () => {
    description("This test exercises the 'meaningful click' heuristic when dispatching clicks on the document node, and requires WebKitTestRunner.");
    if (!window.testRunner)
        return;

    didDispatchNonMeaningfulClickCallback = false;
    testRunner.installDidNotHandleTapAsMeaningfulClickCallback(() => {
        didDispatchNonMeaningfulClickCallback = true;
    });

    didDispatchClick = false;
    document.addEventListener("click", () => {
        didDispatchClick = true;
    });

    await UIHelper.activateAt(100, 100);
    await UIHelper.waitForDoubleTapDelay();
    await new Promise(resolve => shouldBecomeEqual("didDispatchClick", "true", resolve));
    await new Promise(resolve => shouldBecomeEqual("didDispatchNonMeaningfulClickCallback", "true", resolve));

    testRunner.clearTestRunnerCallbacks();
    finishJSTest();
});
</script>
</head>
<body>
    <div id="description"></div>
    <div id="console"></div>
</body>
</html>
